Fix/bug 208 (#148)
* remove list prices in provider

* fix: se cambia el tipo de los porcentajes de payment para que tomen numeros con coma

+++ server/src/main/java/com/agro/feature/payment/domain/Payment.java
@@ -25,9 +25,9 @@ public class Payment {
 
     private Application application;
 
-    private int percentage;
+    private double percentage;
 
-    private int bonusPercentage;
+    private double bonusPercentage;
 
     @ManyToOne
     @JoinColumn(name = "vigentePayments_id")
